#519 Add prison categories to prison. (#187)
Adult male prisons are classified according to the categories of prisoners that they can hold.  There are four categories 'A', 'B', 'C', 'D'.
This commit adds a `var categories: MutableSet<Category>` field to the Prison entity to model that.
A new table, 'prison_category', stores the prison categorisations in the database.
Finally, categories have been added to those adult male prisons that offer Accredited Programmes because this information is required by the hmpps-accredited-programmes-ui service.